backend/requirements.txt
fastapi
pydantic
sqlalchemy
python-dotenv
psycopg2
uvicorn
